Added support for tobii eyetrackers with the tobii research sdk
