+++ pygaze/eyetracker.py
@@ -54,9 +54,9 @@ class EyeTracker(BaseEyeTracker):
 			trackertype = u'dummy'
 
 		# correct wrong input
-		if trackertype not in [u'dumbdummy', u'dummy', u'eyelink', u'smi', u'tobii', u'eyetribe', u'opengaze', u'tobiiglasses']:
+		if trackertype not in [u'dumbdummy', u'dummy', u'eyelink', u'smi', u'eyetribe', u'opengaze', u'tobii', u'tobii-legacy', u'tobiiglasses']:
 			raise Exception( \
-				u"Error in eyetracker.EyeTracker: trackertype '%s' not recognized; it should be one of 'dumbdummy', 'dummy', 'eyelink', 'smi', 'tobii', 'eyetribe', u'opengaze', u'tobiiglasses'" % trackertype)
+				u"Error in eyetracker.EyeTracker: trackertype '%s' not recognized; it should be one of 'dumbdummy', 'dummy', 'eyelink', 'smi', 'eyetribe', 'opengaze', 'tobii', 'tobii-legacy', 'tobiiglasses'" % trackertype)
 
 		# EyeLink
 		if trackertype == u'eyelink':
@@ -76,12 +76,30 @@ class EyeTracker(BaseEyeTracker):
 			# initialize
 			self.__class__.__init__(self, display, **args)
 
+		# Tobii Legacy
+		elif trackertype == u'tobii-legacy':
+			# import libraries
+			from pygaze._eyetracker.libtobiilegacy import TobiiTracker
+			# morph class
+			self.__class__ = TobiiTracker
+			# initialize
+			self.__class__.__init__(self, display, **args)
+
 		# Tobii
 		elif trackertype == u'tobii':
 			# import libraries
-			from pygaze._eyetracker.libtobii import TobiiTracker
+			from pygaze._eyetracker.libtobii import TobiiProTracker
 			# morph class
-			self.__class__ = TobiiTracker
+			self.__class__ = TobiiProTracker
+			# initialize
+			self.__class__.__init__(self, display, **args)
+
+		# Tobii Pro Glasses 2
+		elif trackertype == u'tobiiglasses':
+			# import libraries
+			from pygaze._eyetracker.libtobiiglasses import TobiiGlassesTracker
+			# morph class
+			self.__class__ = TobiiGlassesTracker
 			# initialize
 			self.__class__.__init__(self, display, **args)
 
@@ -120,15 +138,6 @@ class EyeTracker(BaseEyeTracker):
 			self.__class__ = DumbDummy
 			# initialize
 			self.__class__.__init__(self, display)
-
-		# Tobii Pro Glasses 2
-		elif trackertype == u'tobiiglasses':
-			# import libraries
-			from pygaze._eyetracker.libtobiiglasses import TobiiGlassesTracker
-			# morph class
-			self.__class__ = TobiiGlassesTracker
-			# initialize
-			self.__class__.__init__(self, display, **args)
 
 		else:
 			raise Exception( \
